api/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pillow
torch
torchvision
